requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
djangorestframework-simplejwt==4.8.0
certifi==2021.10.8
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
simplejson==3.17.5
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7